create a new interface IUserReviewService with it's class implementation and moved GetUserReviews functionality to it, so that i can schedule a job to update user rate and avoid the circular dependency
