+++ AggarApi/CORE/Services/ReviewService.cs
@@ -219,67 +219,6 @@ namespace CORE.Services
             };
         }
 
-        public async Task<ResponseDto<IEnumerable<SummarizedReviewDto>>> GetUserReviewsAsync(int userId, int pageNo, int pageSize, int maxPageSize = 100)
-        {
-            if(PaginationHelpers.ValidatePaging(pageNo, pageSize, maxPageSize) is string paginationError)
-            {
-                _logger.LogWarning("Invalid pagination parameters: {ErrorMessage}", paginationError);
-                return new ResponseDto<IEnumerable<SummarizedReviewDto>>
-                {
-                    StatusCode = StatusCodes.BadRequest,
-                    Message = paginationError
-                };
-            }
-
-            var userRentalsResponse = await _rentalService.GetRentalsByUserIdAsync(userId, pageNo, pageSize, maxPageSize);
-            if (userRentalsResponse.StatusCode != StatusCodes.OK)
-            {
-                _logger.LogWarning("Failed to retrieve rentals for user {UserId}: {ErrorMessage}",
-                    userId, userRentalsResponse.Message);
-                return new ResponseDto<IEnumerable<SummarizedReviewDto>>
-                {
-                    StatusCode = userRentalsResponse.StatusCode,
-                    Message = userRentalsResponse.Message
-                };
-            }
-            
-            var rentals = userRentalsResponse.Data;
-            if (rentals == null || rentals.Any() == false)
-            {
-                _logger.LogInformation("No rentals found for user {UserId}", userId);
-                return new ResponseDto<IEnumerable<SummarizedReviewDto>>
-                {
-                    StatusCode = StatusCodes.BadRequest,
-                    Message = "No rentals found for this user"
-                };
-            }
-
-            var result = new List<SummarizedReviewDto>();
-            if (rentals.First().Booking.CustomerId == userId)
-            {
-                //get renter reviews on that customer
-                var renterReviewsIds = rentals.Select(r => r.RenterReviewId).ToHashSet();
-                var includes = new List<string> { RenterReviewIncludes.Renter };
-                var reviews = await _unitOfWork.RenterReviews.FindAsync(r => renterReviewsIds.Contains(r.Id), pageNo, pageSize, includes.ToArray());
-                result = _mapper.Map<IEnumerable<SummarizedReviewDto>>(reviews).ToList();
-            }
-            else
-            {
-                //get customer reviews on that renter
-                var customerReviewsIds = rentals.Select(r => r.CustomerReviewId).ToHashSet();
-                var includes = new List<string> { CustomerReviewIncludes.Customer };
-                var reviews = await _unitOfWork.CustomerReviews.FindAsync(r => customerReviewsIds.Contains(r.Id), pageNo, pageSize, includes.ToArray());
-                result = _mapper.Map<IEnumerable<SummarizedReviewDto>>(reviews).ToList();
-            }
-
-            _logger.LogInformation("Successfully retrieved reviews for user {UserId}", userId);
-            return new ResponseDto<IEnumerable<SummarizedReviewDto>>
-            {
-                StatusCode = StatusCodes.OK,
-                Data = result
-            };
-        }
-
         public async Task<ResponseDto<GetReviewDto>> GetReviewAsync(int reviewId)
         {
             _logger.LogInformation("Received request to get review with ID {ReviewId}", reviewId);
